Create a LaTeX component to separate logic
