+++ src/app/components/elements/PageTitle.tsx
@@ -1,10 +1,10 @@
 import React, {ReactElement, useEffect, useRef} from "react";
 import {UncontrolledTooltip} from "reactstrap";
 import {SITE, SITE_SUBJECT, SITE_SUBJECT_TITLE} from "../../services/siteConstants";
-import {TrustedHtml} from "./TrustedHtml";
 import {setMainContentId} from "../../state/actions";
 import {useDispatch, useSelector} from "react-redux";
 import {AppState} from "../../state/reducers";
+import {LaTeX} from "./LaTeX";
 
 export interface PageTitleProps {
     currentPageTitle: string;
@@ -29,7 +29,7 @@ export const PageTitle = ({currentPageTitle, subTitle, help, className, level}: 
     }, [currentPageTitle]);
 
     return <h1 id="main-heading" tabIndex={-1} ref={headerRef} className={`h-title h-secondary${className ? ` ${className}` : ""}`}>
-        <TrustedHtml span html={currentPageTitle} />
+        <LaTeX markup={currentPageTitle} />
         {SITE_SUBJECT === SITE.PHY && level !== undefined && level !== 0 &&
             <span className="float-right h-subtitle">Level {level}</span>}
         {help && <span id="title-help">Help</span>}
