+++ src/app/components/elements/Tabs.tsx
@@ -1,7 +1,7 @@
 import React, {ReactNode, useState} from "react";
 import {Nav, NavItem, NavLink, TabContent, TabPane} from "reactstrap";
-import {TrustedHtml} from "./TrustedHtml";
 import {pauseAllVideos} from "../content/IsaacVideo";
+import {LaTeX} from "./LaTeX";
 
 
 type StringOrTabFunction = string | ((tabTitle: string, tabIndex: number) => string);
@@ -51,7 +51,7 @@ export const Tabs = (props: TabsProps) => {
                         tag="button" type="button" name={tabTitle.replace(" ", "_")}
                         tabIndex={0} className={classes} onClick={() => changeTab(tabIndex)}
                     >
-                        <TrustedHtml html={tabTitle} />
+                        <LaTeX markup={tabTitle} />
                     </NavLink>
                 </NavItem>;
             })}
